fix overflow in rnnocl.cpp

+++ projects/miopen/src/ocl/rnnocl.cpp
@@ -2507,7 +2507,7 @@ void RNNDescriptor::RNNForwardInferencePacked(const Handle& handle,
                                        workSpace,
                                        hx_desc,
                                        hy,
-                                       static_cast<int>(offset) + hid_off +
+                                       offset + hid_off +
                                            static_cast<size_t>(ri) * hy_h +
                                            static_cast<size_t>(use_batch) * hy_stride,
                                        hx_shift + static_cast<size_t>(ri) * hy_n * hy_h +
@@ -2523,7 +2523,7 @@ void RNNDescriptor::RNNForwardInferencePacked(const Handle& handle,
                                        workSpace,
                                        hx_desc,
                                        cy,
-                                       static_cast<int>(offset) + bi * wei_len + ri * hy_h +
+                                       offset + bi * wei_len + ri * hy_h +
                                            use_batch * hy_stride,
                                        hx_shift + static_cast<size_t>(ri) * hy_n * hy_h +
                                            static_cast<size_t>(use_batch) * hy_h);
@@ -3773,8 +3773,8 @@ void RNNDescriptor::RNNForwardTrainingPackedTensors(
                                    reserveSpace,
                                    sp_desc,
                                    reserveSpace,
-                                   static_cast<int>(offset) + 2 * hy_h + ri * wei_len,
-                                   static_cast<int>(offset) + hid_off +
+                                   offset + 2 * hy_h + ri * wei_len,
+                                   offset + hid_off +
                                        static_cast<size_t>(ri) * hy_h +
                                        static_cast<size_t>(nLayers) * batch_n * hy_stride);
                         // Update time
@@ -4035,7 +4035,7 @@ void RNNDescriptor::RNNForwardTrainingPackedTensors(
                                        reserveSpace,
                                        hx_desc,
                                        hy,
-                                       static_cast<int>(offset) + hid_off +
+                                       offset + hid_off +
                                            static_cast<size_t>(ri) * hy_h +
                                            static_cast<size_t>(use_batch) * hy_stride,
                                        hx_shift + static_cast<size_t>(ri) * hy_n * hy_h +
@@ -4051,7 +4051,7 @@ void RNNDescriptor::RNNForwardTrainingPackedTensors(
                                        reserveSpace,
                                        hx_desc,
                                        cy,
-                                       static_cast<int>(offset) + bi * wei_len + ri * hy_h +
+                                       offset + bi * wei_len + ri * hy_h +
                                            use_batch * hy_stride,
                                        hx_shift + static_cast<size_t>(ri) * hy_n * hy_h +
                                            static_cast<size_t>(use_batch) * hy_h);
@@ -4729,7 +4729,7 @@ void RNNDescriptor::RNNBackwardDataPackedTensors(
                                            sp_desc,
                                            workSpace,
                                            pretime_shift + static_cast<size_t>(2) * hy_h,
-                                           static_cast<int>(offset) + ri * wei_len + 2 * hy_h);
+                                           offset + ri * wei_len + 2 * hy_h);
                                 // Update time
                                 profileRNNkernels(handle, 1, ctime);
 
@@ -4771,7 +4771,7 @@ void RNNDescriptor::RNNBackwardDataPackedTensors(
                                 w,
                                 weitime_shift + static_cast<size_t>(ri) * wei_len * uni_stride,
                                 workSpace,
-                                static_cast<int>(offset) + dhd_off + ri * hy_h,
+                                offset + dhd_off + ri * hy_h,
                                 GemmBackend_t::rocblas);
 
                             if(gemm_status != miopenStatusSuccess)
@@ -4795,7 +4795,7 @@ void RNNDescriptor::RNNBackwardDataPackedTensors(
                                            workSpace,
                                            sp_desc,
                                            workSpace,
-                                           static_cast<int>(offset) + ri * wei_len + 2 * hy_h,
+                                           offset + ri * wei_len + 2 * hy_h,
                                            pretime_shift + static_cast<size_t>(2) * hy_h);
                                 // Update time
                                 profileRNNkernels(handle, 1, ctime);
